Added 'EncryptWipeStorage' API to MountService API
The EncryptWipeStorage API is used to create a new ext4 file
system on the userdata partition, instead of the existing one,
and encrypt it.
This as opposed to the way EncryptStorage API works, which
encrypts the existing file system as is ('inplace').
The 'wipe' option is already supported in the underlying Cryptfs
implementation.

Also in this change, new values that can be returned by
'getEncryptionState' API are declared.
These values reflects the state of the MDTP activation, together
with the general encryption state, in case that MDTP is activated.
- ENCRYPTION_STATE_OK_MDTP_ACTIVATED - means that the crypto state is ok,
and MDTP is activated.
- ENCRYPTION_STATE_ERROR_MDTP_ACTIVATED - means that the crypto state is
bad, and MDTP is activated.

Change-Id: Ide628a8cf6499bc2216b08c22479a37133bebb03

Conflicts:
	core/java/android/os/storage/IMountService.java
	services/core/java/com/android/server/MountService.java