Support password input style of times countdown
When password is wrong more than the customized times,
keyguard could show the dialog to suggest to wipe the data.

By default the count is customized to 0, which will disable
this feature. Customize the config_max_unlock_countdown_times
to enable this feature in the operator data package overlay.

Change-Id: I9ba7cd46cb63d447a357783f68854c72cab04c38
CRs-Fixed: 1037633